Change the return type of list_bento, list_model, and refactor common part of yatai,bento client to basecloud
