+++ src/bentoml/_internal/cloud/basecloud.py
@@ -20,6 +20,67 @@ from ..bento import Bento
 from ..bento import BentoStore
 from ..models import Model
 from ..models import ModelStore
+from functools import wraps
+
+
+FILE_CHUNK_SIZE = 100 * 1024 * 1024  # 100Mb
+
+
+class ObjectWrapper(object):
+    def __getattr__(self, name: str) -> t.Any:
+        return getattr(self._wrapped, name)
+
+    def __setattr__(self, name: str, value: t.Any) -> None:
+        return setattr(self._wrapped, name, value)
+
+    def wrapper_getattr(self, name: str):
+        """Actual `self.getattr` rather than self._wrapped.getattr"""
+        return getattr(self, name)
+
+    def wrapper_setattr(self, name: str, value: t.Any) -> None:
+        """Actual `self.setattr` rather than self._wrapped.setattr"""
+        return object.__setattr__(self, name, value)
+
+    def __init__(self, wrapped: t.Any):
+        """
+        Thin wrapper around a given object
+        """
+        self.wrapper_setattr("_wrapped", wrapped)
+
+
+class CallbackIOWrapper(ObjectWrapper):
+    def __init__(
+        self,
+        callback: t.Callable[[int], None],
+        stream: t.BinaryIO,
+        method: t.Literal["read", "write"] = "read",
+    ):
+        """
+        Wrap a given `file`-like object's `read()` or `write()` to report
+        lengths to the given `callback`
+        """
+        super().__init__(stream)
+        func = getattr(stream, method)
+        if method == "write":
+
+            @wraps(func)
+            def write(data: t.Union[bytes, bytearray], *args: t.Any, **kwargs: t.Any):
+                res = func(data, *args, **kwargs)
+                callback(len(data))
+                return res
+
+            self.wrapper_setattr("write", write)
+        elif method == "read":
+
+            @wraps(func)
+            def read(*args: t.Any, **kwargs: t.Any):
+                data = func(*args, **kwargs)
+                callback(len(data))
+                return data
+
+            self.wrapper_setattr("read", read)
+        else:
+            raise KeyError("Can only wrap read/write methods")
 
 
 class BaseCloudClient(ABC):
