+++ src/bentoml/_internal/cloud/bentocloud.py
@@ -8,7 +8,6 @@ import threading
 from pathlib import Path
 from datetime import datetime
 from tempfile import NamedTemporaryFile
-from functools import wraps
 from concurrent.futures import ThreadPoolExecutor
 
 import fs
@@ -45,75 +44,18 @@ from .schemas import CreateModelRepositorySchema
 from .schemas import CompleteMultipartUploadSchema
 from .schemas import PreSignMultipartUploadUrlSchema
 from .basecloud import BaseCloudClient
+from .basecloud import CallbackIOWrapper
+from .basecloud import FILE_CHUNK_SIZE
 from ...exceptions import NotFound
 from ...exceptions import BentoMLException
 from ..configuration.containers import BentoMLContainer
+
+import typing as t
 
 if t.TYPE_CHECKING:
     from concurrent.futures import Future
 
     from rich.progress import TaskID
-
-
-FILE_CHUNK_SIZE = 100 * 1024 * 1024  # 100Mb
-
-
-class ObjectWrapper(object):
-    def __getattr__(self, name: str) -> t.Any:
-        return getattr(self._wrapped, name)
-
-    def __setattr__(self, name: str, value: t.Any) -> None:
-        return setattr(self._wrapped, name, value)
-
-    def wrapper_getattr(self, name: str):
-        """Actual `self.getattr` rather than self._wrapped.getattr"""
-        return getattr(self, name)
-
-    def wrapper_setattr(self, name: str, value: t.Any) -> None:
-        """Actual `self.setattr` rather than self._wrapped.setattr"""
-        return object.__setattr__(self, name, value)
-
-    def __init__(self, wrapped: t.Any):
-        """
-        Thin wrapper around a given object
-        """
-        self.wrapper_setattr("_wrapped", wrapped)
-
-
-class CallbackIOWrapper(ObjectWrapper):
-    def __init__(
-        self,
-        callback: t.Callable[[int], None],
-        stream: t.BinaryIO,
-        method: t.Literal["read", "write"] = "read",
-    ):
-        """
-        Wrap a given `file`-like object's `read()` or `write()` to report
-        lengths to the given `callback`
-        """
-        super().__init__(stream)
-        func = getattr(stream, method)
-        if method == "write":
-
-            @wraps(func)
-            def write(data: t.Union[bytes, bytearray], *args: t.Any, **kwargs: t.Any):
-                res = func(data, *args, **kwargs)
-                callback(len(data))
-                return res
-
-            self.wrapper_setattr("write", write)
-        elif method == "read":
-
-            @wraps(func)
-            def read(*args: t.Any, **kwargs: t.Any):
-                data = func(*args, **kwargs)
-                callback(len(data))
-                return data
-
-            self.wrapper_setattr("read", read)
-        else:
-            raise KeyError("Can only wrap read/write methods")
-
 
 class BentoCloudClient(BaseCloudClient):
     def push_bento(
@@ -1025,14 +967,9 @@ class BentoCloudClient(BaseCloudClient):
         res = yatai_rest_client.get_bentos_list()
         if res is None:
             raise BentoMLException("List bentos request failed")
-
-        res = [
-            {
-                "tag": Tag(bento.repository.name, bento.name),
-                "size": bento.manifest.size_bytes,
-                "creation_time": bento.created_at,
-                "models": bento.manifest.models,
-            }
+        
+        res.items = [
+            bento
             for bento in sorted(res.items, key=lambda x: x.created_at, reverse=True)
         ]
         return res
@@ -1045,13 +982,10 @@ class BentoCloudClient(BaseCloudClient):
         res = yatai_rest_client.get_models_list()
         if res is None:
             raise BentoMLException("List models request failed")
-        res = [
-            {
-                "tag": Tag(model.repository.name, model.name),
-                "size": model.manifest.size_bytes,
-                "creation_time": model.created_at,
-                "module": model.manifest.module,
-            }
+        
+        res.items = [
+            model
             for model in sorted(res.items, key=lambda x: x.created_at, reverse=True)
         ]
         return res
+
